User data API Ready for production

+++ Atlantis.UserData.DAL/UserDAO.cs
@@ -16,7 +16,7 @@ namespace Atlantis.UserData.DAL
             _context = context;
         }
 
-        public User Add(User entity)
+        public virtual User Add(User entity)
         {
             try
             {
@@ -35,13 +35,13 @@ namespace Atlantis.UserData.DAL
                     throw new Exception("User already registered.");
                 }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 throw;
             }
         }
 
-        public List<User> All()
+        public virtual List<User> All()
         {
             try
             {
@@ -71,7 +71,19 @@ namespace Atlantis.UserData.DAL
             {
                 return _context.User.Find(id);
             }
-            catch (Exception ex)
+            catch (Exception)
+            {
+                throw;
+            }
+        }
+
+        public virtual User GetByUserId(string userId)
+        {
+            try
+            {
+                return _context.User.Where(x => x.UserId == userId).FirstOrDefault();
+            }
+            catch (Exception)
             {
                 throw;
             }
@@ -94,13 +106,35 @@ namespace Atlantis.UserData.DAL
             try
             {
                 var usr = _context.User.Find(id);
-                _context.User.Remove(usr);
-                _context.SaveChanges();
+                if (usr != null)
+                {
+                    _context.User.Remove(usr);
+                    _context.SaveChanges();
+                }
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 throw;
             }
+        }
+
+        public virtual bool RemoveByUserId(string userId)
+        {
+            try
+            {
+                var usr = _context.User.Where(x => x.UserId == userId).FirstOrDefault();
+                if (usr != null)
+                {
+                    _context.User.Remove(usr);
+                    _context.SaveChanges();
+                    return true;
+                }
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+            return false;
         }
 
         public User Update(User newEntity)
